irodsapi/requirements.txt
asgiref==3.2.5
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.4
django-environ==0.4.5
django-extensions==2.2.8
globus-sdk==1.9.0
humanize==2.4.0
idna==2.9
josepy==1.3.0
mozilla-django-oidc==1.2.3
prettytable==0.7.2
psycopg2-binary==2.8.4
pycparser==2.20
pycryptodome==3.9.7
PyJWT==1.7.1
pyOpenSSL==19.1.0
pytz==2019.3
requests==2.25.1
six==1.14.0
sqlparse==0.3.1
urllib3==1.26.5
xmlrunner==1.7.7
globus-sdk==1.9.0
jsonschema
python-irodsclient@git+https://github.com/lexis-project/python-irodsclient.git@openid_20210628
django-logging-json
django-cors-headers
django-tus
python-keycloak==0.25.0
celery
django-celery
